Feat:prevent deleting menu items that do not belong to users:
- in menu.py add if statement to deleteMenuItem to check if item belongs to user logged in.
- if item does not belong to user render script to show pop up that lets user know they do not authorization to delete item

+++ menu.py
@@ -316,6 +316,8 @@ def deleteMenuItem(restaurant_id, menu_id):
         return redirect('/login')
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     deleteItem = session.query(MenuItem).filter_by(id=menu_id).one()
+    if login_session['user_id'] != restaurant.user_id:
+        return "<script>function myFunction() {alert('You are not authorized to delete this item');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(deleteItem)
         session.commit()
